refactor(core): rename solver/ to consistency/
The module name now reflects its domain purpose -- consistency checking --
rather than its implementation role. All internal and external import
paths updated.
