+++ dbcop_core/src/consistency/saturation/committed_read.rs
@@ -8,7 +8,7 @@ use crate::history::atomic::types::TransactionId;
 use crate::history::raw::error::Error as NonAtomicError;
 use crate::history::raw::types::{Event, EventId, Session};
 use crate::history::raw::{get_all_writes, get_committed_writes, is_valid_history};
-use crate::solver::error::Error;
+use crate::consistency::error::Error;
 use crate::Consistency;
 
 /// Checks if a valid history is a committed read history.
