+++ dbcop_core/src/history/atomic/types.rs
@@ -13,7 +13,7 @@ use crate::history::raw::error::Error as NonAtomicError;
 use crate::history::raw::types::{Event, EventId};
 use crate::{
     history::raw::{get_all_writes, types::Session},
-    solver::{error::Error, repeatable_read::check_repeatable_read},
+    consistency::{error::Error, repeatable_read::check_repeatable_read},
 };
 use hashbrown::{HashMap, HashSet};
